+++ django_modern_rest/openapi/extractors/__init__.py
@@ -1,4 +1,3 @@
-from django_modern_rest.openapi.extractors import stdlib as stdlib
 from django_modern_rest.openapi.extractors.base import (
     FieldExtractor as FieldExtractor,
 )
