+++ django_test_app/server/apps/controllers/urls.py
@@ -41,11 +41,6 @@ router = Router([
         name='async_parse_headers',
     ),
     path(
-        'negotiation',
-        views.ContentNegotiationController.as_view(),
-        name='negotiation',
-    ),
-    path(
         'constrained-user',
         views.ConstrainedUserController.as_view(),
         name='constrained_user_create',
