+++ django_test_app/server/urls.py
@@ -22,6 +22,7 @@ from django_modern_rest.routing import Router
 from server.apps.controllers import urls as controllers_urls
 from server.apps.middlewares import urls as middleware_urls
 from server.apps.models_example import urls as models_example_urls
+from server.apps.negotiations import urls as negotiations_urls
 from server.apps.openapi.urls import build_spec
 
 router = Router([
@@ -44,6 +45,13 @@ router = Router([
         include(
             (controllers_urls.router.urls, 'controllers'),
             namespace='controllers',
+        ),
+    ),
+    path(
+        'negotiations/',
+        include(
+            (negotiations_urls.router.urls, 'negotiations'),
+            namespace='negotiations',
         ),
     ),
 ])
