+++ tests/test_integration/test_contollers/test_controller_with_negotiation.py
@@ -6,7 +6,7 @@ from django.urls import reverse
 
 from django_modern_rest.negotiation import ContentType
 
-_URL: Final = reverse('api:controllers:negotiation')
+_URL: Final = reverse('api:negotiations:negotiation')
 _XML_DATA = '<root><key>value</key></root>'
